Fix: (FE) Add Carousel Component
1. react-slick을 이용한 이미지 캐루셀 컴포넌트 추가
2. react-slick 로직 수정
