+++ src/components/common/FormToolkit/index.tsx
@@ -3,6 +3,7 @@ import Input from "@src/components/common/FormToolkit/Input";
 import Button from "@src/components/common/FormToolkit/Button";
 import Textarea from "@src/components/common/FormToolkit/Textarea";
 import SinglePhotoInput from "@src/components/common/FormToolkit/SinglePhotoInput";
+import MultiPhotoInput from "@src/components/common/FormToolkit/MultiPhotoInput";
 
 /** 2023/03/15 - Form과 관련된 컴포넌트들 모음 - by 1-blue */
 const FormToolkit = {
@@ -10,6 +11,7 @@ const FormToolkit = {
   Button,
   Textarea,
   SinglePhotoInput,
+  MultiPhotoInput,
 };
 
 export default FormToolkit;
